Handle API cert

+++ dev-analytics-api/kubernetes-deploy.sh
@@ -19,9 +19,17 @@ trap finish EXIT
 rm -rf "$dd"
 mkdir "$dd"
 api_url="api\.${TF_DIR}\.lfanalytics\.io"
+cert=`cat "dev-analytics-api/secrets/ssl-cert.$1.secret"`
+cert="${cert//\//\\\/}"
+cert="${cert//\./\\\.}"
+if [ -z "$cert" ]
+then
+  echo "$0: you need to provide values in dev-analytics-api/secrets/ssl-cert.$1.secret"
+  exit 2
+fi
 cp ~/dev/dev-analytics-api/.circleci/deployments/$API_DIR/* "$dd" || exit 3
 vim --not-a-term -c "%s/image: .*/image: $DOCKER_USER\/dev-analytics-api/g" -c '%s/"bundle", "exec", "rails", "s", "-b", "0\.0\.0\.0"/"\/bin\/sh", "-c", "bundle exec rails db:reset \&\& bundle exec rails s -b 0\.0\.0\.0"/g' -c 'wq!' "${dd}/api.deployment.yml.erb"
-vim --not-a-term -c "%s/external-dns\.alpha\.kubernetes\.io\/hostname: .*/external-dns\.alpha\.kubernetes\.io\/hostname: ${api_url}/g" -c "%s/service.beta.kubernetes.io\/aws-load-balancer.*//g" -c 'wq!' "${dd}/api.deployment.yml.erb"
+vim --not-a-term -c "%s/external-dns\.alpha\.kubernetes\.io\/hostname: .*/external-dns\.alpha\.kubernetes\.io\/hostname: ${api_url}/g" -c "%s/service\.beta\.kubernetes\.io\/aws-load-balancer-ssl-cert: .*/service\.beta\.kubernetes\.io\/aws-load-balancer-ssl-cert: ${cert}/g" -c 'wq!' "${dd}/api.deployment.yml.erb"
 vim --not-a-term -c "%s/image: .*/image: $DOCKER_USER\/dev-analytics-api/g" -c 'wq!' "${dd}/migrate.yml.erb"
 cat dev-analytics-api/sortinghat.partial >> "${dd}/api.deployment.yml.erb"
 cat dev-analytics-api/env.partial >> "${dd}/api.deployment.yml.erb"
